+++ src/post/artist/artist.service.ts
@@ -1,49 +1,54 @@
-import { HttpStatus, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { DataService } from 'src/common/services';
 import { Artist } from 'src/common/interfaces';
-import { isArtistData } from 'src/common/utils';
 import { randomUUID } from 'crypto';
-import { isUUID } from 'class-validator';
 
 @Injectable()
 export class ArtistService {
   constructor(private dataService: DataService) {}
-  create(createArtistDto: Artist) {
-    let s = HttpStatus.CREATED;
-    if (!isArtistData(createArtistDto)) s = HttpStatus.BAD_REQUEST;
+  create(data: Artist) {
     const dto = {
       id: randomUUID({ disableEntropyCache: true }),
-      ...createArtistDto,
+      ...data,
     };
-    return { status: s, body: this.dataService.createArtist(dto) };
+    this.dataService.createArtist(dto);
+    return dto;
   }
 
   findAll() {
-    return { status: HttpStatus.OK, body: this.dataService.getAllArtists() };
+    return this.dataService.getAllArtists();
   }
 
   findOne(id: string) {
-    let s = HttpStatus.OK;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const artist = this.dataService.getArtist(id);
-    if (!artist && s === HttpStatus.OK) s = HttpStatus.NOT_FOUND;
-    return { status: s, body: artist };
+    return this.findAll().find((el) => el.id === id);
   }
 
-  update(id: string, updateArtistDto: Artist) {
-    let s = HttpStatus.OK;
-    if (!isUUID(id, '4') || !isArtistData(updateArtistDto))
-      s = HttpStatus.BAD_REQUEST;
-    const artist = this.dataService.updateArtist(id, updateArtistDto);
-    if (!artist && s === HttpStatus.OK) s = HttpStatus.NOT_FOUND;
-    return { status: s, body: artist };
+  update(id: string, dto: Artist) {
+    const data = this.dataService.getAllArtists();
+    const index = data.findIndex((el) => el.id === id);
+    if (index < 0) return false;
+    const updatedDto = {
+      ...data[index],
+      ...dto,
+    };
+    this.dataService.updateArtist(index, updatedDto);
+    return updatedDto;
   }
 
   remove(id: string) {
-    let s = HttpStatus.NO_CONTENT;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const artist = this.dataService.deleteArtist(id);
-    if (!artist && s === HttpStatus.NO_CONTENT) s = HttpStatus.NOT_FOUND;
-    return { status: s, body: artist };
+    this.dataService.getAllTracks().forEach((track, index) => {
+      if (track.artistId === id) {
+        this.dataService.updateTrack(index, { ...track, artistId: null });
+      }
+    });
+    this.dataService.getAllAlbums().forEach((album, index) => {
+      if (album.artistId === id) {
+        this.dataService.updateAlbum(index, { ...album, artistId: null });
+      }
+    });
+    const data = this.findOne(id);
+    if (!data) return false;
+    this.dataService.deleteArtist(id);
+    return id;
   }
 }
