+++ src/post/favorites/album/album.service.ts
@@ -1,5 +1,4 @@
-import { HttpStatus, Injectable } from '@nestjs/common';
-import { isUUID } from 'class-validator';
+import { Injectable } from '@nestjs/common';
 import { DataService } from 'src/common/services';
 
 @Injectable()
@@ -7,25 +6,19 @@ export class FavAlbumService {
   constructor(private dataService: DataService) {}
 
   create(id: string) {
-    let s = HttpStatus.CREATED;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const createResponse = this.dataService.addAlbumToFav(id);
-    if (!createResponse && s === HttpStatus.CREATED)
-      s = HttpStatus.UNPROCESSABLE_ENTITY;
-    return { status: s, body: createResponse };
+    const isExists = this.dataService.getAllAlbums().find((el) => el.id === id);
+    if (!isExists) return false;
+    this.dataService.addAlbumToFav(id);
+    return id;
   }
 
   findAll() {
-    const allAlbumFavs = this.dataService.getAllAlbumFavs();
-    return { status: HttpStatus.OK, body: allAlbumFavs };
+    return this.dataService.getAllAlbumFavs();
   }
 
   remove(id: string) {
-    let s = HttpStatus.NO_CONTENT;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const deleteResponse = this.dataService.deleteAlbumFav(id);
-    if (!deleteResponse && s === HttpStatus.NO_CONTENT)
-      s = HttpStatus.UNPROCESSABLE_ENTITY;
-    return { status: s, body: deleteResponse };
+    if (!this.findAll().includes(id)) return false;
+    this.dataService.deleteAlbumFav(id);
+    return id;
   }
 }
