+++ src/post/favorites/artist/artist.controller.ts
@@ -1,7 +1,18 @@
-import { Controller, Get, Post, Param, Delete, Res } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Param,
+  Delete,
+  Res,
+  HttpStatus,
+} from '@nestjs/common';
 import { FavArtistService } from './artist.service';
 import { Response } from 'express';
 import { sendResponse } from 'src/common/utils';
+import { isUUID } from 'class-validator';
+
+const target = 'artist';
 
 @Controller()
 export class FavArtistController {
@@ -9,19 +20,25 @@ export class FavArtistController {
 
   @Post(':id')
   create(@Param('id') id: string, @Res() res: Response) {
-    const addResponse = this.artistService.create(id);
-    return sendResponse[addResponse.status](addResponse.body, res);
+    if (!isUUID(id, '4'))
+      return sendResponse[HttpStatus.BAD_REQUEST](res, target);
+    const req = this.artistService.create(id);
+    if (!req) return sendResponse[HttpStatus.UNPROCESSABLE_ENTITY](res, target);
+    return sendResponse[HttpStatus.CREATED](res, req);
   }
 
   @Get()
   findAll(@Res() res: Response) {
-    const allArtistFavs = this.artistService.findAll();
-    return sendResponse[allArtistFavs.status](allArtistFavs.body, res);
+    const req = this.artistService.findAll();
+    return sendResponse[HttpStatus.OK](res, req);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string, @Res() res: Response) {
-    const addResponse = this.artistService.remove(id);
-    return sendResponse[addResponse.status](addResponse.body, res);
+    if (!isUUID(id, '4'))
+      return sendResponse[HttpStatus.BAD_REQUEST](res, target);
+    const req = this.artistService.remove(id);
+    if (!req) return sendResponse[HttpStatus.UNPROCESSABLE_ENTITY](res, target);
+    return sendResponse[HttpStatus.NO_CONTENT](res, req);
   }
 }
