+++ src/post/favorites/track/track.controller.ts
@@ -1,7 +1,18 @@
-import { Controller, Get, Post, Param, Delete, Res } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Param,
+  Delete,
+  Res,
+  HttpStatus,
+} from '@nestjs/common';
 import { FavTrackService } from './track.service';
 import { Response } from 'express';
 import { sendResponse } from 'src/common/utils';
+import { isUUID } from 'class-validator';
+
+const target = 'track';
 
 @Controller()
 export class FavTrackController {
@@ -9,19 +20,25 @@ export class FavTrackController {
 
   @Post(':id')
   create(@Param('id') id: string, @Res() res: Response) {
-    const addResponse = this.trackService.create(id);
-    return sendResponse[addResponse.status](addResponse.body, res);
+    if (!isUUID(id, '4'))
+      return sendResponse[HttpStatus.BAD_REQUEST](res, target);
+    const req = this.trackService.create(id);
+    if (!req) return sendResponse[HttpStatus.UNPROCESSABLE_ENTITY](res, target);
+    return sendResponse[HttpStatus.CREATED](res, req);
   }
 
   @Get()
   findAll(@Res() res: Response) {
-    const allTrackFavs = this.trackService.findAll();
-    return sendResponse[allTrackFavs.status](allTrackFavs.body, res);
+    const req = this.trackService.findAll();
+    return sendResponse[HttpStatus.OK](res, req);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string, @Res() res: Response) {
-    const addResponse = this.trackService.remove(id);
-    return sendResponse[addResponse.status](addResponse.body, res);
+    if (!isUUID(id, '4'))
+      return sendResponse[HttpStatus.BAD_REQUEST](res, target);
+    const req = this.trackService.remove(id);
+    if (!req) return sendResponse[HttpStatus.UNPROCESSABLE_ENTITY](res, target);
+    return sendResponse[HttpStatus.NO_CONTENT](res, req);
   }
 }
